refactor: rename cache hook
