+++ src/shared/hooks/useCache.ts
